requirements.txt
aiofiles>=0.8.0
aiohttp>=3.8.0
anthropic
asyncio-mqtt
bcrypt>=4.0.1
cryptography>=41.0.0
docling
google-generativeai>=0.3.0
mcp-agent
mcp-server-git
nest_asyncio
openai>=1.0.0
openrouter>=0.1.0
pathlib2
psycopg2-binary>=2.9.0
PyJWT>=2.8.0
PyPDF2>=2.0.0
pyotp>=2.9.0
qrcode>=7.4.0
reportlab>=3.5.0
requests>=2.31.0
SQLAlchemy>=2.0.0
streamlit>=1.29.0
streamlit-authenticator>=0.3.0
